requirements.txt
scikit-learn>=1.3
pandas>=2.0
numpy>=1.24
optuna>=3.4
mlflow>=2.10
bentoml>=1.1
dvc>=3.40
pydantic>=2.0
pyyaml>=6.0
matplotlib>=3.7
seaborn>=0.13
shap>=0.44
folium>=0.16
optuna-integration[mlflow]>=3.6
